+++ gui/slick/js/configNotifications.js
@@ -377,68 +377,73 @@ $(document).ready(function(){
 	});
 
 	$('#testPushbullet').click(function () {
-		var pushbullet_api = $.trim($('#pushbullet_api').val());
-		if (!pushbullet_api) {
+		var pushbullet_access_token = $.trim($('#pushbullet_access_token').val());
+		var pushbullet_device_iden = $('#pushbullet_device_iden').val();
+		if (!pushbullet_access_token) {
 			$('#testPushbullet-result').html('Please fill out the necessary fields above.');
-			$('#pushbullet_api').addClass('warning');
+			$('#pushbullet_access_token').addClass('warning');
 			return;
 		}
-		$('#pushbullet_api').removeClass('warning');
+		$('#pushbullet_access_token').removeClass('warning');
 		$(this).prop('disabled', true);
 		$('#testPushbullet-result').html(loading);
-		$.get(sbRoot + '/home/testPushbullet', {'api': pushbullet_api})
+		$.get(sbRoot + '/home/testPushbullet', {'accessToken': pushbullet_access_token, 'device_iden': pushbullet_device_iden})
 			.done(function (data) {
 				$('#testPushbullet-result').html(data);
 				$('#testPushbullet').prop('disabled', false);
 			});
 	});
 
-	function get_pushbullet_devices(msg){
-
-		if(msg){
+	function get_pushbullet_devices (msg) {
+		var pushbullet_access_token = $.trim($('#pushbullet_access_token').val());
+		if (!pushbullet_access_token) {
+			$('#testPushbullet-result').html('Please fill out the necessary fields above.');
+			$('#pushbullet_access_token').addClass('warning');
+			return;
+		}
+		$(this).prop("disabled", true);
+		if (msg) {
 			$('#testPushbullet-result').html(loading);
 		}
-
-		var pushbullet_api = $("#pushbullet_api").val();
-
-		if(!pushbullet_api) {
-			$('#testPushbullet-result').html("You didn't supply a Pushbullet api key");
-			$("#pushbullet_api").focus();
-			return false;
-		}
-
-		var current_pushbullet_device = $("#pushbullet_device").val();
-		$.get(sbRoot + "/home/getPushbulletDevices", {'api': pushbullet_api},
-			function (data) {
-				var devices = jQuery.parseJSON(data).devices;
-				$("#pushbullet_device_list").html('');
-				for (var i = 0; i < devices.length; i++) {
-					if(devices[i].active == true) {
-						if(current_pushbullet_device == devices[i].iden) {
-							$("#pushbullet_device_list").append('<option value="'+devices[i].iden+'" selected>' + devices[i].nickname + '</option>')
-						} else {
-							$("#pushbullet_device_list").append('<option value="'+devices[i].iden+'">' + devices[i].nickname + '</option>')
+		var current_pushbullet_device = $('#pushbullet_device_iden').val();
+		$.get(sbRoot + '/home/getPushbulletDevices', {'accessToken': pushbullet_access_token})
+			.done(function (data) {
+				var devices = jQuery.parseJSON(data || '{}').devices;
+				$('#pushbullet_device_list').html('');
+				// add default option to send to all devices
+				$('#pushbullet_device_list').append('<option value="" selected="selected">-- All Devices --</option>');
+				if (devices) {
+					for (var i = 0; i < devices.length; i++) {
+						// only list active device targets
+						if (devices[i].active == true) {
+							// if a device in the list matches our current iden, select it
+							if (current_pushbullet_device == devices[i].iden) {
+								$('#pushbullet_device_list').append('<option value="' + devices[i].iden + '" selected="selected">' + devices[i].manufacturer + ' ' + devices[i].nickname + '</option>');
+							} else {
+								$('#pushbullet_device_list').append('<option value="' + devices[i].iden + '">' + devices[i].manufacturer + ' ' + devices[i].nickname + '</option>');
+							}
 						}
 					}
 				}
-				if(msg) {
+				$('#getPushbulletDevices').prop('disabled', false);
+				if (msg) {
 					$('#testPushbullet-result').html(msg);
 				}
-			}
-		);
+			});
 
-		$("#pushbullet_device_list").change(function(){
-			$("#pushbullet_device").val($("#pushbullet_device_list").val());
-			$('#testPushbullet-result').html("Don't forget to save your new pushbullet settings.");
+		$('#pushbullet_device_list').change(function () {
+			$('#pushbullet_device_iden').val($('#pushbullet_device_list').val());
+			$('#testPushbullet-result').html('Don\'t forget to save your new Pushbullet settings.');
 		});
-	};
+	}
 
-	$('#getPushbulletDevices').click(function(){
-		get_pushbullet_devices("Device list updated. Please choose a device to push to.");
+	$('#getPushbulletDevices').click(function () {
+		get_pushbullet_devices('Device list updated. Select specific device to use.');
 	});
 
-	// we have to call this function on dom ready to create the devices select
-	get_pushbullet_devices();
+	if ($('#use_pushbullet').prop('checked')) {
+		get_pushbullet_devices();
+	}
 
 	$('#email_show').change(function () {
 		var key = parseInt($('#email_show').val(), 10);